IS_project_source_code/source_code/requirements.txt
numpy
tensorflow
opencv-python
tk
imutils
Pillow